test: added account declare and deploy test on already existing account on sepolia
