+++ tests/account_katana.rs
@@ -11,25 +11,13 @@ use beerus::{
         TxnHash,
     },
 };
+use common::constants::{
+    CLASS_HASH, COMPILED_ACCOUNT_CONTRACT_V2, COMPILED_ACCOUNT_CONTRACT_V3,
+    CONTRACT_ADDRESS, DECLARE_ACCOUNT_V2, DECLARE_ACCOUNT_V3, SENDER_ADDRESS,
+};
 use common::katana::Katana;
 
 mod common;
-
-pub const COMPILED_ACCOUNT_CONTRACT_V2: &str =
-    include_str!("./clob/compiled_account_contract_v2.txt");
-pub const COMPILED_ACCOUNT_CONTRACT_V3: &str =
-    include_str!("./clob/compiled_account_contract_v3.txt");
-pub const DECLARE_ACCOUNT_V2: &str =
-    include_str!("./clob/declare_account_v2.txt");
-pub const DECLARE_ACCOUNT_V3: &str =
-    include_str!("./clob/declare_account_v3.txt");
-
-pub const CLASS_HASH: &str =
-    "0x6b46f84b1bbb779e588a9c5f577907c3dfb66e6b13cf4c4f480d4fb1677c2ba";
-pub const CONTRACT_ADDRESS: &str =
-    "0x5f177e96c32bcc09607d9e2adf1864b8a8dc54c7fb5a283afac58f2b6ddc7da";
-pub const SENDER_ADDRESS: &str =
-    "0x6162896d1d7ab204c7ccac6dd5f8e9e7c25ecd5ae4fcb4ad32e57786bb46e03";
 
 async fn setup() -> (Katana, Client<Http>) {
     let katana = Katana::init("http://127.0.0.1:0").await.unwrap();
