+++ tests/common/mod.rs
@@ -1,5 +1,6 @@
 pub mod constants;
 pub mod ctx;
+pub mod dummies;
 pub mod err;
 pub mod katana;
 pub mod matchers;
